+++ locales/__init__.py
@@ -1,0 +1,3 @@
+from . import text
+from . import buttons
+from . import other
