+++ locales/buttons.py
