+++ src/__init__.py
@@ -1,8 +1,6 @@
-from . import other
-from . import buttons
-from . import text
-
-from . import filters
-from . import objects
 from . import system
+from . import regex
+from . import instances
+from . import parsers
+from . import filters
 from . import utils
